+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+### Fix
+- Fix setup fatal error should not trigger a redirect
 
 ## [2.1.0] - 2018-06-14
 ### Add